requirements.txt
google-generativeai
streamlit
python-dotenv
langchain_core
langchain_community